Use minified jQuery by default.
While I personally prefer developing with an unminified jQuery (as
it makes debugging easier) I understand that most people don't need that.

So the onus is now on more advanced developers to s/.min.js/.js/
Though they're safe with the build script which will switch the refs back to the min.

And everyone else will default to min'd js, regardless of if they use the
build script.

:)