Remove placeholder link styles
The placeholder link styles don't add much value but do introduce
potentially unwanted specificity issues. By specifying `a:visited` it
forces subsequent styles for components (like buttons) to have a greater
specificity than this selector if they are to avoid the color changing
once it has been visited.

A similar problem is also evident if a user builds their link styles
upon the project styles (rather than editing the ones included).

This was also an issue in Normalize.css and was fixed in the same way.