Change the outdated browser prompt classname
While the outdated browser prompt currently links to `browsehappy.com`,
users are free and even encouraged to substitute the link with their own
page or site of preference.

This commit changes the `browsehappy` specific classname with the more
agnostic `browserupgrade` in order to make the link substitution
easier, but more importantly to hint to the users that they don't have
to stick with the currently default provided `browsehappy.com` link.

Ref: h5bp/html5-boilerplate#1582
     h5bp/html5-boilerplate#1608

Close h5bp/html5-boilerplate#1608